Move `rand_circuit` and `rand_gate` in `random_objects`
